+++ src/core/server/player/index.ts
@@ -3,6 +3,7 @@ export * as currency from '@AthenaServer/player/currency';
 export * as emit from '@AthenaServer/player/emit';
 export * as events from '@AthenaServer/player/events';
 export * as inventory from '@AthenaServer/player/inventory';
+export * as permission from '@AthenaServer/player/permissions';
 export * as safe from '@AthenaServer/player/safe';
 export * as set from '@AthenaServer/player/setter';
 export * as sync from '@AthenaServer/player/sync';