+++ src/core/server/systems/permission.ts
@@ -115,11 +115,7 @@ const InternalFunctions = {
         }
 
         const data = Athena.document[dataName].get(player);
-        if (typeof data === 'undefined') {
-            return false;
-        }
-
-        if (typeof data.permissions === 'undefined') {
+        if (typeof data === 'undefined' || typeof data.permissions === 'undefined') {
             return false;
         }
 
@@ -149,7 +145,7 @@ const InternalFunctions = {
         }
 
         const data = Athena.document[dataName].get(player);
-        if (typeof data === 'undefined') {
+        if (typeof data === 'undefined' || typeof data.permissions === 'undefined') {
             return false;
         }
 
@@ -157,10 +153,6 @@ const InternalFunctions = {
         // Should also check this before checking player perms if they haven't had any perms added yet.
         if (perms.length <= 0) {
             return true;
-        }
-
-        if (typeof data.permissions === 'undefined') {
-            return false;
         }
 
         if (data.permissions.length <= 0) {
@@ -197,7 +189,7 @@ const InternalFunctions = {
         }
 
         const data = Athena.document[dataName].get(player);
-        if (typeof data === 'undefined') {
+        if (typeof data === 'undefined' || typeof data.permissions === 'undefined') {
             return false;
         }
 
@@ -205,10 +197,6 @@ const InternalFunctions = {
         // Should also check this before checking player perms if they haven't had any perms added yet.
         if (perms.length <= 0) {
             return true;
-        }
-
-        if (typeof data.permissions === 'undefined') {
-            return false;
         }
 
         if (data.permissions.length <= 0) {
@@ -458,5 +446,3 @@ export default {
     getAll,
     removeAll,
 };
-
-