+++ src/core/server/vehicle/index.ts
@@ -6,6 +6,7 @@ export * as despawn from './despawn';
 export * as events from './events';
 export * as get from './get';
 export * as ownership from './ownership';
+export * as permissions from './permissions';
 export * as shared from './shared';
 export * as spawn from './spawn';
 export * as tempVehicles from './tempVehicles';